+++ apps/app/src/features/cms/server/routes/apiv3/index.ts
@@ -6,7 +6,7 @@ import Crowi from '~/server/crowi';
 module.exports = (crowi: Crowi): Router => {
   const router = express.Router();
 
-  router.use('/namespace', require('./namespace')(crowi));
+  router.use('/namespace', require('./cms-namespace')(crowi));
 
   return router;
 };
